+++ tests/Manga.AcceptanceTests/V1/SunnyDayTests.cs
@@ -47,7 +47,7 @@ namespace Manga.AcceptanceTests.V1
             string result = await client.GetStringAsync("/api/v1/Accounts/" + accountId);
         }
 
-        private async Task<Tuple<string, string>> Register(double initialAmount)
+        private async Task<Tuple<string, string>> Register(decimal initialAmount)
         {
             var client = _factory.CreateClient();
             var register = new
@@ -75,7 +75,7 @@ namespace Manga.AcceptanceTests.V1
             return new Tuple<string, string>(customerId, accountId);
         }
 
-        private async Task Deposit(string account, double amount)
+        private async Task Deposit(string account, decimal amount)
         {
             var client = _factory.CreateClient();
             var json = new
@@ -93,7 +93,7 @@ namespace Manga.AcceptanceTests.V1
             response.EnsureSuccessStatusCode();
         }
 
-        private async Task Withdraw(string account, double amount)
+        private async Task Withdraw(string account, decimal amount)
         {
             var client = _factory.CreateClient();
             var json = new
