+++ tests/Manga.UnitTests/UseCaseTests/CloseAccount/PositiveDataSetup.cs
@@ -2,13 +2,13 @@ namespace Manga.UnitTests.UseCasesTests.CloseAccount
 {
     using Xunit;
 
-    internal sealed class PositiveDataSetup : TheoryData<double>
+    internal sealed class PositiveDataSetup : TheoryData<decimal>
     {
         public PositiveDataSetup()
         {
-            Add(0.5);
-            Add(100);
-            Add(200);
+            Add(0.5M);
+            Add(100M);
+            Add(200M);
         }
     }
 }
